// backend/src/constants/enums.ts
export enum LoanType {
  GIVEN = "GIVEN",
  TAKEN = "TAKEN",
}

export enum LoanStatus {
  ACTIVE = "ACTIVE",
  PARTIALLY_PAID = "PARTIALLY_PAID",
  COMPLETED = "COMPLETED",
  OVERDUE = "OVERDUE",
}

export enum PaymentType {
  RECEIVED = "RECEIVED",
  PAID = "PAID",
}

export enum PaymentMethod {
  CASH = "CASH",
  BANK = "BANK",
  JAZZCASH = "JAZZCASH",
  EASYPAISA = "EASYPAISA",
  OTHER = "OTHER",
}